@@ -1,4 +1,4 @@
-const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
+import {getDefaultConfig, mergeConfig} from '@react-native/metro-config';
 
 /**
  * Metro configuration
