@@ -1,6 +1,7 @@
+// eslint-disable-next-line no-undef
 module.exports = {
-  presets: ['module:@react-native/babel-preset'],
-  plugins: [
-    'react-native-reanimated/plugin',
-  ],
+    presets: ['module:@react-native/babel-preset'],
+    plugins: [
+        'react-native-reanimated/plugin',
+    ],
 };
