@@ -1,9 +1,10 @@
+// eslint-disable-next-line no-undef
 module.exports = {
     dependencies: {
-      'react-native-vector-icons': {
-        platforms: {
-          ios: null,
+        'react-native-vector-icons': {
+            platforms: {
+                ios: null,
+            },
         },
-      },
     },
-  };
+};
