@@ -1,9 +1,9 @@
-import LoginScreen from '../screens/LoginScreen';
+import {LoginScreen} from '../screens/LoginScreen';
 
 export const authenticationRoutes = {
-  loginScreen: {
-    name: 'login',
-    component: LoginScreen,
-    options: {title: 'Welcome', headerShown: false },
-  },
+    loginScreen: {
+        name: 'login',
+        component: LoginScreen,
+        options: {title: 'Welcome', headerShown: false },
+    },
 };
