@@ -1,5 +1,5 @@
 import {authenticationRoutes} from './authenticationRoutes';
 
 export const routes = {
-  ...authenticationRoutes,
+    ...authenticationRoutes,
 };
